Create Scala 2 veneers for src/main/scala/chisel3/util/BitPat.scala
